Silence Xcode Errors
Mostly initialization order, unused functions, missing virtual destructors, missing overrides.

One instance of returning a pointer to a temporary variable was fixed in the StateHandler class.

+++ bdsp_dsp/Math/bdsp_MultiDimensionalLookup.h
@@ -145,7 +145,7 @@ namespace bdsp
 			}
 
 		private:
-			int getIndex(size_t x, size_t y) const
+			size_t getIndex(size_t x, size_t y) const
 			{
 				return x + y * dimX;
 			}
